Internal change
PiperOrigin-RevId: 236341203
